Created a new file named utils and all methods from main have been moved there except main and argparse.
Added new tests and some of them have been updated.
